+++ package.json
@@ -15,29 +15,27 @@
     "lint": "node_modules/.bin/tslint --project tsconfig.json --type-check --force"
   },
   "dependencies": {
+    "echarts": "^4.4.0",
+    "formidable": "^1.2.1",
     "iview": "^3.3.0",
     "iview-weapp": "^2.0.0",
-    "echarts": "^4.4.0",
-    "tinymce": "^5.0.7",
-    "vue": "^2.5.2",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.0.1",
-    "formidable": "^1.2.1",
     "js-yaml": "^3.11.0",
     "jshtml": "^0.3.0",
     "koa2-cors": "^2.0.6",
     "log4js": "^4.0.2",
     "lowdb": "^1.0.0",
     "opn": "^4.0.2",
-    "ora": "^1.2.0"
+    "ora": "^1.2.0",
+    "tinymce": "^5.0.7",
+    "vue": "^2.5.2",
+    "vue-router": "^3.0.1",
+    "vuex": "^3.0.1"
   },
   "devDependencies": {
-    "@types/node": "^12.7.2",
-    "@vue/cli-service-global": "^4.0.4",
-    "connect-timeout": "^1.9.0",
-    "express-http-proxy": "^1.6.0",
     "@tinymce/tinymce-vue": "^2.0.0",
     "@types/echarts": "^4.4.0",
+    "@types/node": "^12.7.2",
+    "@vue/cli-service-global": "^4.0.4",
     "autoprefixer": "^6.7.2",
     "axios": "^0.19.0",
     "babel-core": "^6.22.1",
@@ -47,26 +45,16 @@
     "babel-preset-stage-2": "^6.22.0",
     "babel-register": "^6.22.0",
     "chalk": "^1.1.3",
-    "ts-loader": "^2.0.3",
-    "tslint": "^5.9.1",
-    "typescript": "^2.8.1",
-    "url-loader": "^0.5.8",
-    "vuex-class": "^0.3.0",
-    "vue-loader": "^11.3.4",
-    "vue-property-decorator": "^6.0.0",
-    "vue-style-loader": "^2.0.5",
-    "vue-template-compiler": "^2.2.6",
-    "vueify": "^9.4.1",
-    "vuex-module-decorators": "^0.9.9",
-    "yarn": "^1.19.1",
     "clean-webpack-plugin": "v1.0.0",
     "compression": "^1.7.4",
     "compression-webpack-plugin": "^1.1.11",
     "connect-history-api-fallback": "^1.3.0",
+    "connect-timeout": "^1.9.0",
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.28.0",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.14.1",
+    "express-http-proxy": "^1.6.0",
     "extract-text-webpack-plugin": "^2.0.0",
     "file-loader": "^0.11.1",
     "friendly-errors-webpack-plugin": "^1.1.3",
@@ -76,29 +64,42 @@
     "jade": "^1.11.0",
     "jade-loader": "^0.8.0",
     "optimize-css-assets-webpack-plugin": "^1.3.0",
+    "pug": "^2.0.4",
+    "pug-filters": "^3.1.1",
+    "pug-loader": "^2.4.0",
     "rimraf": "^2.6.0",
     "semver": "^5.3.0",
     "shelljs": "^0.7.6",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2",
-    "pug": "^2.0.4",
-    "pug-filters": "^3.1.1",
-    "pug-loader": "^2.4.0",
+    "ts-loader": "^2.0.3",
+    "tslint": "^5.9.1",
+    "typescript": "^2.8.1",
+    "url-loader": "^0.5.8",
     "vue-cli": "^2.9.6",
+    "vue-loader": "^11.3.4",
+    "vue-property-decorator": "^6.0.0",
+    "vue-router-invoke-webpack-plugin": "^0.3.0",
+    "vue-style-loader": "^2.0.5",
+    "vue-template-compiler": "^2.2.6",
+    "vueify": "^9.4.1",
+    "vuex-class": "^0.3.0",
+    "vuex-module-decorators": "^0.9.9",
     "webpack": "^2.3.3",
     "webpack-bundle-analyzer": "^3.6.0",
     "webpack-cli": "^3.3.9",
     "webpack-dev-middleware": "^1.10.0",
     "webpack-dev-server": "^3.8.2",
     "webpack-hot-middleware": "^2.18.0",
-    "webpack-merge": "^4.1.0"
+    "webpack-merge": "^4.1.0",
+    "yarn": "^1.19.1"
   },
   "config": {
     "forge": "./config/forge.config.js"
   },
   "proxyServer": {
     "port": 6666,
-    "host": "192.168.1.18"
+    "host": "39.108.103.150"
   },
   "engines": {
     "node": ">= 4.0.0",
