+++ package/webpack.build.conf.js
@@ -1,130 +1,138 @@
-var path = require('path')
-var utils = require('./utils')
-var webpack = require('webpack')
-var config = require('../config')
-var merge = require('webpack-merge')
-var baseWebpackConfig = require('./webpack.base.conf')
-var HtmlWebpackPlugin = require('html-webpack-plugin')
-var FriendlyErrorsPlugin = require('friendly-errors-webpack-plugin')
-var ExtractTextPlugin = require('extract-text-webpack-plugin')
-var OptimizeCSSPlugin = require('optimize-css-assets-webpack-plugin')
-var CleanWebpackPlugin = require('clean-webpack-plugin')
-var CopyWebpackPlugin = require('copy-webpack-plugin')
+var path = require('path');
+var utils = require('./utils');
+var webpack = require('webpack');
+var config = require('../config');
+var merge = require('webpack-merge');
+var baseWebpackConfig = require('./webpack.base.conf');
+var HtmlWebpackPlugin = require('html-webpack-plugin');
+var FriendlyErrorsPlugin = require('friendly-errors-webpack-plugin');
+var ExtractTextPlugin = require('extract-text-webpack-plugin');
+var OptimizeCSSPlugin = require('optimize-css-assets-webpack-plugin');
+var CleanWebpackPlugin = require('clean-webpack-plugin');
+var CopyWebpackPlugin = require('copy-webpack-plugin');
+const VueRouterInvokeWebpackPlugin = require('vue-router-invoke-webpack-plugin');
 
 // add hot-reload related code to entry chunks
 Object.keys(baseWebpackConfig.entry).forEach(function (name) {
-  baseWebpackConfig.entry[name] = ['./package/dev-client'].concat(baseWebpackConfig.entry[name])
-})
+    baseWebpackConfig.entry[name] = ['./package/dev-client'].concat(baseWebpackConfig.entry[name]);
+});
 
 var webpackConfig = merge(baseWebpackConfig, {
-  module: {
-    rules: utils.styleLoaders({ sourceMap: config.build.cssSourceMap })
-  },
-  output: {
-    path: config.build.assetsRoot
-  },
-  // cheap-module-eval-source-map is faster for development
-  devtool: config.build.cssSourceMap ? '#source-map' : false,
-  plugins: [
-    new webpack.DefinePlugin({
-      'process.env': config.build.env
-    }),
-    new CleanWebpackPlugin([config.build.assetsRoot]),
-    // https://github.com/glenjamin/webpack-hot-middleware#installation--usage
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.NoEmitOnErrorsPlugin(),
-    // https://github.com/ampedandwired/html-webpack-plugin
-    new HtmlWebpackPlugin({
-        filename: config.build.index,
-        template: 'src/views/index.pug',
-        inject: true,
-        minify: {
-            removeComments: true,
-            collapseWhitespace: true,
-            removeAttributeQuotes: true
-            // more options:
-            // https://github.com/kangax/html-minifier#options-quick-reference
-        },
-        // necessary to consistently work with multiple chunks via CommonsChunkPlugin
-        chunksSortMode: 'dependency'
-    }),
-    new FriendlyErrorsPlugin(),
-    // split vendor js into its own file
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
-      minChunks: function (module, count) {
-          // any required modules inside node_modules are extracted to vendor
-          return (
-              module.resource &&
-              /\.js$/.test(module.resource) &&
-              module.resource.indexOf(
-                  path.join(__dirname, '../node_modules')
-              ) === 0
-          )
-      }
-    }),
-    // extract webpack runtime and module manifest to its own file in order to
-    // prevent vendor hash from being updated whenever app bundle is updated
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'manifest',
-      chunks: ['vendor']
-    }),
+    module: {
+        rules: utils.styleLoaders({ sourceMap: config.build.cssSourceMap })
+    },
+    output: {
+        path: config.build.assetsRoot
+    },
+    // cheap-module-eval-source-map is faster for development
+    devtool: config.build.cssSourceMap ? '#source-map' : false,
+    plugins: [
+        new webpack.DefinePlugin({
+            'process.env': config.build.env
+        }),
+        new CleanWebpackPlugin([config.build.assetsRoot]),
+        new VueRouterInvokeWebpackPlugin({
+            'dir' : 'src/components',
+            'alias' : '../../components',
+            'routerDir' : 'src/router',
+            'language' : 'typescript',
+            'mode' : 'hash'
+        }),
+        // https://github.com/glenjamin/webpack-hot-middleware#installation--usage
+        new webpack.HotModuleReplacementPlugin(),
+        new webpack.NoEmitOnErrorsPlugin(),
+        // https://github.com/ampedandwired/html-webpack-plugin
+        new HtmlWebpackPlugin({
+            filename: config.build.index,
+            template: 'src/views/index.pug',
+            inject: true,
+            minify: {
+                removeComments: true,
+                collapseWhitespace: true,
+                removeAttributeQuotes: true
+                // more options:
+                // https://github.com/kangax/html-minifier#options-quick-reference
+            },
+            // necessary to consistently work with multiple chunks via CommonsChunkPlugin
+            chunksSortMode: 'dependency'
+        }),
+        new FriendlyErrorsPlugin(),
+        // split vendor js into its own file
+        new webpack.optimize.CommonsChunkPlugin({
+            name: 'vendor',
+            minChunks: function (module, count) {
+            // any required modules inside node_modules are extracted to vendor
+                return (
+                    module.resource &&
+                    /\.js$/.test(module.resource) &&
+                    module.resource.indexOf(
+                        path.join(__dirname, '../node_modules')
+                    ) === 0
+                );
+            }
+        }),
+        // extract webpack runtime and module manifest to its own file in order to
+        // prevent vendor hash from being updated whenever app bundle is updated
+        new webpack.optimize.CommonsChunkPlugin({
+            name: 'manifest',
+            chunks: ['vendor']
+        }),
 
-    // extract css into its own file
-    new ExtractTextPlugin({
-      filename: utils.assetsPath('css/[name].[contenthash].css')
-    }),
+        // extract css into its own file
+        new ExtractTextPlugin({
+            filename: utils.assetsPath('css/[name].[contenthash].css')
+        }),
 
-    // Compress extracted CSS. We are using this plugin so that possible
-    // duplicated CSS from different components can be deduped.
-    new OptimizeCSSPlugin({
-      cssProcessorOptions: {
-          safe: true
-      }
-    }),
-    // copy custom static assets
-    new CopyWebpackPlugin([
-        {
-            from: path.resolve(__dirname, '../src/assets/tinymce/langs'),
-            to: config.build.assetsRoot + '/langs',
-            ignore: ['.*']
-        },
-        {
-            from: path.resolve(__dirname, '../src/assets/tinymce/skins'),
-            to: config.build.assetsRoot + '/skins',
-            ignore: ['.*']
-        },
-        {
-            from: path.resolve(__dirname, '../src/assets/tinymce/plugins'),
-            to: config.build.assetsRoot + '/plugins',
-            ignore: ['.*']
-        },
-        {
-            from: path.resolve(__dirname, '../src/assets/css'),
-            to: config.build.assetsRoot + '/css',
-            ignore: ['.*']
-        }
-    ])
-  ]
+        // Compress extracted CSS. We are using this plugin so that possible
+        // duplicated CSS from different components can be deduped.
+        new OptimizeCSSPlugin({
+            cssProcessorOptions: {
+                safe: true
+            }
+        }),
+        // copy custom static assets
+        new CopyWebpackPlugin([
+            {
+                from: path.resolve(__dirname, '../src/assets/tinymce/langs'),
+                to: config.build.assetsRoot + '/langs',
+                ignore: ['.*']
+            },
+            {
+                from: path.resolve(__dirname, '../src/assets/tinymce/skins'),
+                to: config.build.assetsRoot + '/skins',
+                ignore: ['.*']
+            },
+            {
+                from: path.resolve(__dirname, '../src/assets/tinymce/plugins'),
+                to: config.build.assetsRoot + '/plugins',
+                ignore: ['.*']
+            },
+            {
+                from: path.resolve(__dirname, '../src/assets/css'),
+                to: config.build.assetsRoot + '/css',
+                ignore: ['.*']
+            }
+        ])
+    ]
 });
 
 
 if (config.build.productionGzip) {
-  var CompressionWebpackPlugin = require('compression-webpack-plugin')
+    var CompressionWebpackPlugin = require('compression-webpack-plugin');
 
-  webpackConfig.plugins.push(
-      new CompressionWebpackPlugin({
-        asset: '[path].gz[query]',
-        algorithm: 'gzip',
-        test: new RegExp(
-            '\\.(' +
-            config.build.productionGzipExtensions.join('|') +
-            ')$'
-        ),
-        threshold: 10240,
-        minRatio: 0.8
-      })
-  )
+    webpackConfig.plugins.push(
+        new CompressionWebpackPlugin({
+            asset: '[path].gz[query]',
+            algorithm: 'gzip',
+            test: new RegExp(
+                '\\.(' +
+                config.build.productionGzipExtensions.join('|') +
+                ')$'
+            ),
+            threshold: 10240,
+            minRatio: 0.8
+        })
+    );
 }
 
 module.exports = webpackConfig;
