+++ src/components/Nav/Contributive/Index.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
     import "@/assets/css/common.css";
-    import TurnWorkStore from '../store/modules/TurnWorkStore';
+    import ContributiveStore from '../../../store/modules/ContributiveStore';
     import { Component, Vue, Prop, Model} from 'vue-property-decorator';
     import { getModule } from 'vuex-module-decorators';
 
@@ -20,16 +20,28 @@
             }
         }
     })
-    export default class TurnWork extends Vue {
+    export default class Contributive extends Vue {
 
         private store: any;
         constructor() {
             super();
-            this.store = getModule(TurnWorkStore)
+            this.store = getModule(ContributiveStore)
         }
 
         @Model('isCollapsed', { type: Boolean }) private isCollapsed !: boolean;
 
+        rowClassName (row, index) : string {
+
+            if(index == 0) {
+                return 'table-header'
+            }
+
+            return '';
+        }
+        mounted() {
+            this.store.search()
+            console.log('mounted');
+        }
         private options!: any;
         getMenus() : any {
             if(this.options) return this.options;
@@ -43,9 +55,11 @@
             return this.store.columns;
         }
         getData() : any{
-            return this.store.data;
+            return this.store.contributive;
         }
+
+
     }
 </script>
-<style scoped src="@/styles/turnWork.css" />
-<template lang="pug" src="@/views/turnWork.pug" />
+<style scoped src="@/styles/contributive.css" />
+<template lang="pug" src="@/views/contributive.pug" />
