+++ src/components/Nav/Cultivate/Index.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
     import "@/assets/css/common.css";
-    import ProjectStore from '../store/modules/ProjectStore';
+    import CultivateStore from '../../../store/modules/CultivateStore';
     import { Component, Vue, Prop, Model} from 'vue-property-decorator';
     import { getModule } from 'vuex-module-decorators';
 
@@ -20,12 +20,12 @@
             }
         }
     })
-    export default class Project extends Vue {
+    export default class Cultivate extends Vue {
 
         private store: any;
         constructor() {
             super();
-            this.store = getModule(ProjectStore)
+            this.store = getModule(CultivateStore)
         }
 
         @Model('isCollapsed', { type: Boolean }) private isCollapsed !: boolean;
@@ -43,8 +43,9 @@
         getMenus() : any {
             if(this.options) return this.options;
             this.options = [
-                { value: '离场', key: '0' },
-                { value: '在场', key: '1' }
+                { value: '待学习', key: '0' },
+                { value: '学习中', key: '1' },
+                { value: '学习完成', key: '2' }
             ];
             return this.options;
         }
@@ -58,5 +59,5 @@
 
     }
 </script>
-<style scoped src="@/styles/project.css" />
-<template lang="pug" src="@/views/project.pug" />
+<style scoped src="@/styles/cultivate.css" />
+<template lang="pug" src="@/views/cultivate.pug" />
