+++ src/components/Nav/ExaminationPaper/Index.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
     import "@/assets/css/common.css";
-    import LecturerStore from '../store/modules/LecturerStore';
+    import ExaminationPaperStore from '../../../store/modules/ExaminationPaperStore';
     import { Component, Vue, Prop, Model} from 'vue-property-decorator';
     import { getModule } from 'vuex-module-decorators';
 
@@ -20,21 +20,19 @@
             }
         }
     })
-    export default class Lecturer extends Vue {
+    export default class ExaminationPaper extends Vue {
 
         private store: any;
         constructor() {
             super();
-            this.store = getModule(LecturerStore)
+            this.store = getModule(ExaminationPaperStore)
         }
-        getLecturer() : any{
-            return this.store.lecturer;
-        }
+
         @Model('isCollapsed', { type: Boolean }) private isCollapsed !: boolean;
-
-
-
+        getPaper() : any{
+            return this.store.paper;
+        }
     }
 </script>
-<style scoped src="@/styles/lecturer.css" />
-<template lang="pug" src="@/views/lecturer.pug" />
+<style scoped src="@/styles/examinationPaper.css" />
+<template lang="pug" src="@/views/examinationPaper.pug" />
