+++ src/components/Nav/Home/Index.vue
