+++ src/components/Nav/Unit/Index.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
     import "@/assets/css/common.css";
-    import ContributiveStore from '../store/modules/ContributiveStore';
+    import UnitStore from '../../../store/modules/UnitStore';
     import { Component, Vue, Prop, Model} from 'vue-property-decorator';
     import { getModule } from 'vuex-module-decorators';
 
@@ -20,12 +20,12 @@
             }
         }
     })
-    export default class Contributive extends Vue {
+    export default class Unit extends Vue {
 
         private store: any;
         constructor() {
             super();
-            this.store = getModule(ContributiveStore)
+            this.store = getModule(UnitStore)
         }
 
         @Model('isCollapsed', { type: Boolean }) private isCollapsed !: boolean;
@@ -38,10 +38,7 @@
 
             return '';
         }
-        mounted() {
-            this.store.search()
-            console.log('mounted');
-        }
+
         private options!: any;
         getMenus() : any {
             if(this.options) return this.options;
@@ -55,11 +52,11 @@
             return this.store.columns;
         }
         getData() : any{
-            return this.store.contributive;
+            return this.store.data;
         }
 
 
     }
 </script>
-<style scoped src="@/styles/contributive.css" />
-<template lang="pug" src="@/views/contributive.pug" />
+<style scoped src="@/styles/unit.css" />
+<template lang="pug" src="@/views/unit.pug" />
