+++ src/main.ts
@@ -3,7 +3,7 @@
 
 import Vue from 'vue'
 import App from './App.vue'
-import router from './router'
+import router from './router/.invoke/router'
 
 import iView from 'iview'
 import 'iview/dist/styles/iview.css'
